requirements.txt
aiohttp==3.7.3
aiokafka==0.7.0
aioschedule==0.5.2
confluent-kafka==1.6.0
fastapi==0.63.0
pendulum==2.1.2
python-dotenv==0.15.0
uvicorn==0.13.3